requirements.txt
# Data processing
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
requests>=2.31.0
scipy>=1.11.0

# Visualization and analysis
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Statistical analysis
statsmodels>=0.14.0


# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
tiktoken>=0.5.1
peft>=0.13.0
accelerate>=0.34.0


# Utilities
tqdm>=4.65.0
psutil>=5.9.5

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0

# Optional dependencies
python-dotenv>=1.0.0
streamlit>=1.33.0